using size_t instead of int64_t, -Wsign-compare

+++ r/src/recordbatch.cpp
@@ -151,7 +151,7 @@ std::shared_ptr<arrow::RecordBatch> ipc___ReadRecordBatch__InputStream__Schema(
 arrow::Status check_consistent_array_size(
     const std::vector<std::shared_ptr<arrow::Array>>& arrays, int64_t* num_rows) {
   *num_rows = arrays[0]->length();
-  for (int64_t i = 1; i < arrays.size(); i++) {
+  for (size_t i = 1; i < arrays.size(); i++) {
     if (arrays[i]->length() != *num_rows) {
       return arrow::Status::Invalid("All arrays must have the same length");
     }
